Create greatest_common_divisor
For this problem, I found the GCD of an array by first identifying the largest and smallest numbers. I then divided the larger number by the smaller one and replaced it with the remainder. I continued this process until the remainder became zero. The last non-zero number left is the GCD of the array.